Separate out synchronize labels, save annotations and pixelmap updates
Too often updates were triggered that were not required for the change made.
These changes enforce the following:
- Only update the pixelmap if a superpixel is painted, border opacity is
changed, or label color is changed.
- Only update the config file when a label name, label color, border opacity,
or exclusions list is changed.
- Only synchronize labels when the label name or label color is changed
- Only save annotations when a label value (name, color) is changed,
superpixel(s) are painted, or review is made.
